requirements.txt
# Core dependencies
streamlit==1.54.0
python-dotenv==1.0.0

# LangChain and related (Updated to current installed versions)
langchain==1.2.9
langchain-community==0.4.1
langchain-core==1.2.9
langchain-ollama==1.0.1
langchain-text-splitters==1.1.0

# Vector store and embeddings
chromadb==1.0.20
sentence-transformers==5.2.2

# Document loaders
pypdf==6.6.2
pymupdf==1.23.8
python-docx==1.1.0
unstructured==0.11.2
unstructured[local-inference]==0.11.2
python-magic==0.4.27
python-magic-bin==0.4.14; platform_system == "Windows"

# Data processing
pandas==2.1.3
openpyxl==3.1.2

# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Azure services
azure-cognitiveservices-speech==1.33.0

# Audio processing
# LLM
ollama==0.6.0

# PDF generation (courtroom transcript)
reportlab>=4.0.4

# Utilities
tqdm==4.66.1
pydantic==2.5.0
typing-extensions==4.8.0

# Optional but recommended
tiktoken==0.5.1
huggingface-hub==0.19.4

# Upload & voice (dynamic input)
PyMuPDF>=1.23.0
sounddevice>=0.4.6
faster-whisper>=0.10.0
scipy>=1.11.0